Add options to search request to include topics or types (#59)
This adds two optional parameters to the /bags/search endpoint, "fillTypes" and "fillTopics", that will cause the search results to include all of the message types and/or topics for all of the returned messages.  Normally these are omitted because the grid in the GUI doesn't need them and deserializing all of that information is slow, but they can be useful for other scripts that use that endpoint.