Fix user being logged out after streaming video
The service call that streamed video was never returning, and that would
cause an internal error in the web server that would invalidate the user's
session and force them to log in again.

This also adds support for configuring the LDAP server through environment
variables that can be set on the Docker container.

Signed-off-by: P. J. Reed <preed@swri.org>