Make it easier to change which subdirectory is used for serving
This adds a BAGDB_PATH environment variable to the Docker container that
will control the path used by the internal Tomcat server to serve the
Bag Database.

This is primarily intended for making it easier to integrate together
with a reverse proxy, so the docker-compose example that uses Traefik
has been updated to also serve the Bag Database from a subdirectory.