Fix running scripts w/ multiple bags with the same name
If multiple bags were in different paths but had the same filename, they
would all be mounted at the same point inside the Docker container used
to run scripts, making it so that only one of them could actually be processed.
This adds the full path to the bag to its mount point to ensure it is unique.

Fixed #130

Signed-off-by: P. J. Reed <preed@swri.org>